requirements.txt
numpy
matplotlib
ipython[all]
scikit-image
torch
torchvision
progressbar2